LayoutTests/streams/reference-implementation/bad-strategies-expected.txt


PASS Readable stream: throwing strategy.size getter 
PASS Readable stream: throwing strategy.size method 
PASS Readable stream: throwing strategy.highWaterMark getter 
PASS Readable stream: invalid strategy.highWaterMark 
PASS Readable stream: invalid strategy.size return value 
FAIL Writable stream: throwing strategy.size getter assert_throws: construction should re-throw the error function "function () {
        new WritableStream({}, {
          ..." threw object "ReferenceError: Can't find variable: WritableStream" ("ReferenceError") expected object "Error: a unique string" ("Error")
FAIL Writable stream: throwing strategy.size method Can't find variable: WritableStream
FAIL Writable stream: invalid strategy.size return value Can't find variable: WritableStream
FAIL Writable stream: throwing strategy.highWaterMark getter assert_throws: construction should re-throw the error function "function () {
        new WritableStream({}, {
          ..." threw object "ReferenceError: Can't find variable: WritableStream" ("ReferenceError") expected object "Error: a unique string" ("Error")
FAIL Writable stream: invalid strategy.highWaterMark assert_throws: construction should throw a RangeError for -1 function "function () {
            new WritableStream({}, {
      ..." threw object "ReferenceError: Can't find variable: WritableStream" ("ReferenceError") expected object "RangeError" ("RangeError")
